Move patches to their own directory.
