+++ lib/cext/patches/websocket_driver_patches.rb
@@ -1,0 +1,22 @@
+# Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved. This
+# code is released under a tri EPL/GPL/LGPL license. You can use it,
+# redistribute it and/or modify it under the terms of the:
+#
+# Eclipse Public License version 1.0
+# GNU General Public License version 2
+# GNU Lesser General Public License version 2.1
+
+class WebsocketDriverPatches < CommonPatches
+
+  PATCHES = {
+    gem: 'websocket-driver',
+    patches: {
+      'websocket_mask.c' => [
+        {
+          match: /(VALUE .*?)\s+= Qnil;/,
+          replacement: '\1;'
+        }
+      ]
+    },
+  }
+end
