--- lib/cext/websocket_driver_patches.rb
@@ -1,14 +1,0 @@
-class WebsocketDriverPatches < CommonPatches
-
-  PATCHES = {
-    gem: 'websocket-driver',
-    patches: {
-      'websocket_mask.c' => [
-        {
-          match: /(VALUE .*?)\s+= Qnil;/,
-          replacement: '\1;'
-        }
-      ]
-    },
-  }
-end
