+++ src/main/c/Makefile
@@ -16,7 +16,7 @@ RUBY_LIBRARY := $(ROOT)/lib/cext/ruby.su
 RUBY_HEADERS := $(wildcard $(ROOT)/lib/cext/*.h) $(wildcard $(ROOT)/lib/cext/ruby/*.h)
 RBCONFIG := $(ROOT)/lib/truffle/rbconfig.rb
 RBCONFIG_MKMF := $(ROOT)/lib/truffle/rbconfig-for-mkmf.rb
-PREPROCESS := $(ROOT)/lib/cext/*.rb
+PREPROCESS := $(ROOT)/lib/cext/preprocess.rb $(wildcard $(ROOT)/lib/cext/patches/*.rb)
 MKMF := $(ROOT)/lib/mri/mkmf.rb
 BASIC_EXTCONF_DEPS := $(TRUFFLE_POSIX) $(RUBY_HEADERS) $(RBCONFIG) $(RBCONFIG_MKMF) $(PREPROCESS) $(MKMF)
 # These 2 libraries are needed for try_link() in extconf.rb
